implemented profile page frontend/backend,  implemented objects for the demographics/voterinfo and issues pages to save values between pages, implemented a check for when user tries logging in without having survey completed, implemented app bar on Home page that contains the PoGo logo, made nav bar neater, some small bugfixes.
Issues page rating bars now keep the user's rating between page switches. I did not implement saving the answers from the Demographics and VoterInfo pages yet as I need to see what Rifat has changed with those pages.

A user object is created through a new page called homeLoadingPage. This page just displays a circular loading progress widget while user info is being pulled from the database. The page switches once the object is instantiated. When the survey backend is working, objects to hold the survey info will have to be instantiated in this page as well.

The first version of the user profile is done. It currently displays the user's full name, email, phone number, and address. It will have to be edited in the future when the backend for the survey is done, when we decided to implement ballot linking through social media, and when we decided to implement AWS storage for uploading and viewing custom profile pictures.