Minor changes to userprofile, voter guide and editperosnalinfo pages
I added labels for the text boxes on the editpersonalinfopage.dart.  Also, a minor problem that prevented the app from saving the changed address was resolved. I added a text line to the User profile's personal information section to display the address. A function that looks through the address variable, locates the state initial, and replaces it was added in Voter guide. Made sure the user's name was displayed, and based on the user's State initial, I added it so selecting "request an absentee ballot" would direct them to the website for that State's absentee ballot requesting site.