had to move remove filter button to ballot page to fix another bug.
When the remove filter button was on Podium page and was tapped, the initial 3 candidate cards would be invisible. They would only appear after navigating to a different page and back. I moved the remove filter button to the ballot page to fix this issue. For now, the button appears on the top right of the ballot page, above the local dropdown, and only when a filter is active.